application/backend/pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = "Backend"
authors = [
    "Stella Parker <116223316+whoIsStella@users.noreply.github.com>"
]

package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.2.2,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
argon2-cffi = "^23.1.0"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"
gunicorn = "^23.0.0"
django-filter = "25.1"
python-dotenv = "^1.1.1"
channels = "^4.2.2"
channels-redis = "^4.2.1"
daphne = "^4.2.1"
whitenoise = "^6.9.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"